Remove caching decorator from VarBytesColumn reader.
My best guess is that issue #386 involves users sharing readers between threads, which is not supported.
That's the only way I can see the cache trying to delete the same thing twice?
I think the error in the cache is masking the real problem, that users are sharing readers.
I'm going to remove the cache because it probably has a very small performance impact and the error is misleading.

Fixes issue #386.