Fix reporting of total count in FilterCollector, based on PR #63 by Jannon Frank.
Set the 'collector' attribute of the results objects to be the FilterCollector.
Removed some code in FilterCollector.get_count() that subtracts the filtered_count from the child's results,
because for collectors that compute the count, the FilterCollector has already only told them to collect on valid filtered items.

Also a minor PEP8 fix in collectors.py.